+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## Version 1.0.0-alpha
+- Stripped down to core features and settings
+- Added polling mode. This is a more cpu-intensive workaround, butfixes several issues:
+  - **Linux is now supported**
+  - Allows dragging electron-generated files into sinkholes
+  - 
+
 ## Version 0.1.1-alpha
 
 - Better logging method
@@ -9,7 +16,7 @@
 
 - Added support `pointer-events` mode  
   This is now the default mode and gets used in the demo.
-- Added support for holes (experimental)
+- Added support for sinkholes (experimental)
 - `blacklist` and `whitelist` mode are now deprecated
 
 ## Version 0.0.1